Rewrite Dockerfile for arm32 Linux samples (#1006)
The samples build is failing for arm32 Linux leg because the build agents have an architecture of aarch64 and the multi-arch manifest description doesn't have a matching architecture for the 2.2 SDK and runtime images that we reference. To fix this, we eventually want to make use of the --platform option on the Docker CLI to explicitly set what platform we want to get but that's only available in experimental mode. To workaround this, we are going to modify the Dockerfile dynamically so it uses the platform-specific tag for the image instead of the 2.2 flag.