Change URL path for get-drops-version script (#1125)
In order to allow us to target releases for the master branch of .NET Core SDK for the update-dependencies-from-drop pipeline, we need to update the URL that is called by the script. The master branch doesn't use "release" in the path. So we'll baseline it to support master and then for other branches, we'll prepend "release/" in the pipeline variable.