Define common engineering infra folder
Consolidates all the infrastructure files in the repo to a single top-level folder: eng. The goal of these changes is to enable a simple copy between .NET Docker-related repos so that they can make use of the same infrastructure. Currently, the files that are shared amongst repos are spread out across several folders and contain slightly variations between repos. The intention is to simply be able to xcopy a common folder between the repos to get them updated.