Add presence check for procps on Debian-based images
On Debian-based distributions, the procps package is required for
the "dotnet watch" command to work accordingly.

This has introduced a regression in the official sdk:5.0 image
recently.

This commit adds a test that ensures this package is present on
Debian-based distribution (we're assuming an image is Debian based
if the "dpkg" command is present).

Issue: https://github.com/dotnet/dotnet-docker/issues/2396