Filter out gpg-pubkey
The gpg-pubkey is not a relevant package to be included in the manifest. It represents a Mariner public key that is automatically installed from /etc/pki/rpm-gpg when you install packages for the first time. It shows up in rpm queries because of this but it's not an RPM package.