Update Azure Artifacts sample for Windows
The Windows sample for Azure Artifacts either got stale or never worked as specified, as the credential provider install did so as the wrong user, preventing dotnet restore from succeeding.

Updating the sample to use an explicit temp directory as Windows Nano Server doesn't have a good location for unelevated temp files. This is done in the build layer so therefore doesn't end up in the final image that contains the published build.